csv2glozz: Restore timestamp based ids
Glozz disregards the ids you assign to annotations and systematically
uses author_creation-date, so we might as well match its behaviour.

Unfortunately this means we no longer get the same result if we run the
script more than once.

In principle, it might be possible to use a decimal representation of
the hash (with lots of digits) for stability instead, but maybe we
shouldn't tempt fate.

NB: if we want to go back to the version of csv2glozz before my
adventures in it, I think r1095 is the one we want, or in my Git
mirror, commit e57cc351cc1721df91c97a1bddc80c02d95ff44d

git-svn-id: https://wwwsecu.irit.fr/svn/Stac@1125 2e4a9ccd-9a2a-43a5-8ce1-c04977dfc52d