csv2glozz: Fix off-by-one regression from last fix
This happens in cases where we have more than one segment
in the same turn:

  p: do& re &m
  01234 5678 9

Last fix was svn:r1959 or git:ce33b8bd7.

There's a couple of things going on here. First, a couple of facts I had
not understood about Glozz: (i) that it is 0-indexed and not 1-indexed
(we see indices of 1 because each .ac file starts with ' '), and (ii)
the rightmost index should not be the position of the rightmost
character but 1 plus (ie. left + length; like a loop boundary).

I'd gotten the left+length part right, but was wrongly following up by
also adding +1 to advance the left boundary of each segment. This error
compounds when you have a whole chain of segments, turning an off-by-one
into on off-by-two, off-by-three...

BUT there's another thing going on, which is that the original script
seems to have its successive left boundaries shifted 1 over from where
they normally would be (compare OK/UGLY 're' with GOOD). I *think* this
is because the original code wrongly assumes that each successive
segment starts with a space (so it would get [m] wrong).  To account for
this, I now nudge the left boundaries over however many whitespace
characters we have in front the segment.

            BAD   | OK/UGLY | GOOD
         +--------+---------+------
 [do]    |   3-5  | 3-5     | 3-5
 [ re ]  |   6-9  | 5-9     | 6-9
 [m]     |  10-11 | 9-10    | 9-10
         +--------+---------+------

git-svn-id: https://wwwsecu.irit.fr/svn/Stac@2011 2e4a9ccd-9a2a-43a5-8ce1-c04977dfc52d