stac-util rewrite: almost a no-op
This reads Glozz files and saves them back again (in the output dir).
This is useful as a precursor to other commands like stac-util move
or stac-util merge-dialogue because it normalises the Glozz XML to
how the script will write it (silly things like indentation, spaces
before '/>', order of features).

By using rewrite, you get to separate the diffs that arise from this
from the diffs that come from your changes proper.

git-svn-id: https://wwwsecu.irit.fr/svn/Stac@2111 2e4a9ccd-9a2a-43a5-8ce1-c04977dfc52d