stac-util filter --not-id
There's some strange bug trying to compose filters, where if
you have one filter, regex matching works fine and returns
None if it fails to match; but if you have two filters, it
returns a zero-length match.

The strange thing about this is is that the two regexes used
are completely unrelated to my awareness.

Anyway --not-id is good enough

git-svn-id: https://wwwsecu.irit.fr/svn/Stac@2174 2e4a9ccd-9a2a-43a5-8ce1-c04977dfc52d