Using corenlp in STAC parser pipeline.
We've always been able to do this, but hesitated because the
corenlp startup time would have been probhitive. Always wanted
to look into some sort of server based solution, something that
would let us start corenlp once and then just make calls to it,
but never got around to it.

So we'll just pay the startup cost for now and worry about it
later. Worst case is that we have to disable the features.

git-svn-id: https://wwwsecu.irit.fr/svn/Stac@2396 2e4a9ccd-9a2a-43a5-8ce1-c04977dfc52d