Feature extraction overhaul (again!)
Hopefully this will make feature extraction easier to maintain.

First: rel-info can now print out a list of all the features
implemented, along with a descriptive message for each one.  The aim is
to get a better handle on the ever evolving question of "which features
have we implemented?"

Second: from the UI perspective the script now requires subcommands
(rel-info extract: feature extraction, rel-info list to list features)

Third: there's a bit less duplication between the following tasks:

- declaring the list of features
- assigning them an order in the CSV file
- implementing the extraction for them
- [NEW] printing help messages for the features
- [NEW] grouping the features into related themes (also needed for
  sensible help text)

Fourth: a bit less code sprawl as we try to move the feature
definitions and help text closer to the implementation.

The situation isn't yet entirely satisfactory, but I think we are
slowly creeping towards a sensible story for it.

git-svn-id: https://wwwsecu.irit.fr/svn/Stac@2585 2e4a9ccd-9a2a-43a5-8ce1-c04977dfc52d