feature extraction: fix control flow efficiency
Oops. We had some code which works a bit like this:

for edu1:
   for edu2:
      1. compute pair vector
      2. continue if not the same dialogue
      3. break if edu split > window

But this arrangement does a bit too much work; you're computing pair
vectors you don't really need to know about (2 can be determined
without the vectors).  A better arrangement is to swap them so that
you get

for edu1:
   for edu2:
      2. continue if not the same dialogue
      1. compute pair vector
      3. break if edu split > window (still depends on 1)

In fact, since the EDUs are sorted, you know that if you're not
same dialogue, subsequent EDUs in the inner cycle are also going
to be out of dialogue, so you can just break:

for edu1:
   for edu2:
      2. break if not the same dialogue
      1. compute pair vector
      3. break if edu split > window

git-svn-id: https://wwwsecu.irit.fr/svn/Stac@2586 2e4a9ccd-9a2a-43a5-8ce1-c04977dfc52d