feature extraction: cache features for single EDUs
Feature extraction for EDU pairs tends to involve reusing the
single EDUs:

x1 | x2 x3 x4 x5 x6 x7
   x2 | x3 x4 x5 x6 x7 x8

The idea here is that if we cache the single EDU features (x1..x7 in the
first step) then they will be available for the next step (x2..x8).

The cache forms a sliding window requiring constant memory (we always
delete the element from the outer loop, so x1 from the first step, x2
in the second step), so the number of items that can be in the cache
is constrained by the window size.

The cache could also come in handy if we want to compute features based
not just on pairs of EDUs but on intermediary ones. For example, if the
pair we have in mind is (x1,x7), we might be interested to know about
features in x2..x6. Recomputng those would be expensive

git-svn-id: https://wwwsecu.irit.fr/svn/Stac@2587 2e4a9ccd-9a2a-43a5-8ce1-c04977dfc52d