MAINT simplify local config
The aim here is to have a less opaque config.
A lot of the opacity is in navigating the mess of
auto-combination logic.

But really sometimes you have to accept some copy and paste programming
for clarity. It's about finding the right places to accept repetition.
Note that we've ditched code for expanding structured learners, but it
shouldn't be too hard to reconstruct it.

I'm much happier with this